% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA24204172.R
\name{AR1_covariance_matrix}
\alias{AR1_covariance_matrix}
\title{A correlation matrix generator using R}
\usage{
AR1_covariance_matrix(size, rho)
}
\arguments{
\item{size}{the number of variance}

\item{rho}{the correlation parameter}
}
\value{
A correlation matrix
}
\description{
generate a AR correlation matrix
}
\examples{
\dontrun{
    Cov <- AR1_covariance_matrix(size,rho)
    pheatmap(Cov,cluster_rows = FALSE, cluster_cols = FALSE,cellwidth = 2, cellheight = 2 )
}
}
