% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA24204172.R
\name{banded}
\alias{banded}
\title{A banding operator using R}
\usage{
banded(X, k)
}
\arguments{
\item{X}{matrix}

\item{k}{bandwidth}
}
\value{
A banded matrix
}
\description{
band a matrix
}
\examples{
\dontrun{
    data_banded <- banded(data,k)
}
}
