% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA24204172.R
\name{generate_symmetric_positive_definite_matrix}
\alias{generate_symmetric_positive_definite_matrix}
\title{A symmetric positive definite matrix generator using R}
\usage{
generate_symmetric_positive_definite_matrix(p)
}
\arguments{
\item{p}{dim}
}
\value{
A symmetric positive definite matrix
}
\description{
generate a random symmetric positive definite matrix
}
\examples{
\dontrun{
    Cov_permuted <- permute(Cov,k)
}
}
