% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{naive_estimator}
\alias{naive_estimator}
\title{A estimator using Rcpp}
\usage{
naive_estimator(Cov, p, q)
}
\arguments{
\item{Cov}{covariance matrix}

\item{p}{dim1}

\item{q}{dim2}
}
\value{
solution
}
\description{
A estimator using Rcpp
}
\examples{
\dontrun{
result <- naive_estimator(Cov,p,q)
}
}
