% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA24204172.R
\name{permute}
\alias{permute}
\title{A reshape operator using R}
\usage{
permute(mat, A_dim1, A_dim2, B_dim1, B_dim2)
}
\arguments{
\item{mat}{matrix}

\item{A_dim1}{dim1}

\item{A_dim2}{dim2}

\item{B_dim1}{dim3}

\item{B_dim2}{dim4}
}
\value{
A reshaped matrix
}
\description{
reshapes a matrix's blocks
}
\examples{
\dontrun{
    Cov_permuted <- permute(Cov,k)
}
}
