% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA24204172.R
\name{propose_estimator}
\alias{propose_estimator}
\title{proposed estimator using R}
\usage{
propose_estimator(Cov, p, q, lambda, maxiter)
}
\arguments{
\item{Cov}{covariance matrix}

\item{p}{dim1}

\item{q}{dim2}

\item{lambda}{penalty weight}

\item{maxiter}{run times}
}
\value{
solution
}
\description{
solve covariance matrix
}
\examples{
\dontrun{
   result <- propose_estimator(Cov,p,q,maxiter)
}
}
