% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA24204172.R
\name{simulate_data}
\alias{simulate_data}
\title{A simulated data generator using R}
\usage{
simulate_data(n, p, q, sigma)
}
\arguments{
\item{n}{the number of simulated data}

\item{p}{the row number}

\item{q}{the column number}

\item{sigma}{covariance matrix}
}
\value{
A simulated data matrix
}
\description{
generate  simulated data follow multivariate normal distribution
}
\examples{
\dontrun{
    simulated_data <- simulate_data(n_samples, p_size, q_size, sigma)
}
}
