% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{naive_estimator}
\alias{naive_estimator}
\title{A Gibbs sampler using Rcpp}
\usage{
naive_estimator(Cov, p, q)
}
\arguments{
\item{Cov}{the number of samples}

\item{p}{the number of between-sample random numbers}

\item{q}{111}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
rnC <- gibbsC(100,10)
par(mfrow=c(2,1));
plot(rnC[,1],type='l')
plot(rnC[,2],type='l')
}
}
